sstable: refactor test code for reuse
