+++ sstable/testdata/reader/iter
@@ -1,5 +1,8 @@
 build
-a:1=A,b:2=B,c:3=C,d:4=D
+a.SET.1:A
+b.SET.2:B
+c.SET.3:C
+d.SET.4:D
 ----
 
 iter
